refactor: give run transactions direct ownership
